requirements.txt
aiohttp-socks==0.8.0
aiohttp==3.8.4
aiorpcX==0.22.1
aiosignal==1.3.1
amqp==5.1.1
aniso8601==9.0.1
anyio==3.6.2
apturl==0.5.2
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
arrow==1.2.3
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
automium-web==0.1.1
automium==0.2.6
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bitstring==4.0.2
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels-redis==4.1.0
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
constantly==15.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
crypto==1.4.1
cryptography==40.0.2
cupshelpers==1.0
cycler==0.11.0
daphne==4.0.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro-info==1.1build1
distro==1.7.0
dj-database-url==1.2.0
django-allauth==0.54.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-filter==23.1
django-timezone-field==5.0
Django==4.2
djangorestframework==3.14.0
dnspython==2.3.0
docker==6.0.1
drf-yasg==1.21.5
duplicity==0.8.21
Electrum==4.4.3
executing==1.2.0
factory-boy==3.2.1
Faker==16.7.0
fasteners==0.14.1
fastjsonschema==2.17.1
filelock==3.9.0
flake8==6.0.0
Flask==2.2.3
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.17.3
googleapis-common-protos==1.59.0
graphene-django==3.0.1
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
inflection==0.5.1
ipykernel==6.23.1
ipython-genutils==0.2.0
ipython==8.13.2
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-console==6.6.3
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter-server-terminals==0.4.4
jupyter-server==2.6.0
jupyter==1.0.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.22.1
jupyterlab-widgets==3.0.7
jupyterlab==4.0.2
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
legacy==0.1.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
matplotlib==3.7.1
mccabe==0.7.0
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.5
multidict==6.0.4
mutagen==1.45.1
mysql-connector-python==8.0.33
Naked==0.1.32
natsort==8.0.2
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook-shim==0.2.3
notebook==6.5.4
numpy==1.24.1
oauthlib==3.2.2
olefile==0.46
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.0.2
pipenv==2022.12.19
platformdirs==2.6.2
prometheus-client==0.17.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pyflakes==3.0.1
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pypng==0.20220715.0
PyQt5-sip==12.9.1
PyQt5==5.15.6
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-crontab==2.7.1
python-dateutil==2.8.2
python-debian==0.1.43ubuntu1
python-dotenv==1.0.0
python-json-logger==2.0.7
python-socks==2.3.0
python-telegram-bot==20.2
python3-openid==3.2.0
pytz==2023.3
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
QDarkStyle==3.1
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
rabbitmq==0.2.0
redis==4.5.5
reportlab==3.6.8
requests-oauthlib==1.3.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.21
SecretStorage==3.3.1
Send2Trash==1.8.2
service-identity==21.1.0
setuptools==59.6.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
stevedore==5.0.0
systemd-python==234
telegram==0.0.1
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.3.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
usb-creator==0.3.7
vine==5.0.0
virtualenv-clone==0.5.7
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==9.1
Werkzeug==2.2.3
wheel==0.37.1
whitenoise==6.3.0
widgetsnbextension==4.0.7
wrapt==1.14.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
xlrd==2.0.1
yarl==1.9.2
yt-dlp==2022.4.8
zipp==1.0.0
zope.interface==6.0
